+++ Bootstrap.php
@@ -96,11 +96,18 @@ class Bootstrap implements BootstrapInterface
         if (!Yii::$app instanceof \yii\console\Application)
             $app->getRequest()->parsers[] = ['application/json' => 'yii\web\JsonParser'];
 
-        // Configure options component
-        $app->setComponents([
-            'api' => [
-                'class' => API::className()
-            ]
-        ]);
+        // Configure API-auth component
+        $request = new \yii\web\Request(['url' => parse_url(Yii::$app->request->getUrl(), PHP_URL_PATH)]);
+        if(preg_match('/^\/api\/?+/is', $request->url)) {
+            $module->controllerNamespace = 'wdmg\api\controllers\api';
+            $app->setComponents([
+                'user' => [
+                    'class' => '\yii\web\User',
+                    'identityClass' => 'wdmg\api\models\API',
+                    'enableAutoLogin' => false,
+                    'enableSession' => false,
+                ]
+            ]);
+        }
     }
 }