+++ messages/ru-RU/app/modules/api.php
@@ -1,7 +1,10 @@
 <?php
 
 return [
-    'API' => "API"
+    'API' => "API",
+
+    'This method not allowed for API.' => "Этот метод не разрешен для API.",
+
 ];
 
 ?>